Refactor logo styles into base folder
