+++ client/src/Styles/bases/General.jsx
